ignores files in file loop

+++ script.py
@@ -1,17 +1,17 @@
 import os, fnmatch, shutil
-import string
 
 result = []
 print('Checking files...')
 path = '/Users/isaactennant/Downloads/'
-dest = '/Users/isaactennant/Desktop/'
+dest = '/Users/isaactennant/Library/Application Support/Sports Interactive/Football Manager 2022/'
 
-for root, dirs, files in os.walk(path):
-    """for file in files:
+for root, dirs, files in os.walk(path, True):
+    for file in files:
         if (fnmatch.fnmatch(file, '*.fmf')):
-            print('file match: ' + file)
-            shutil.move(os.path.join(path, file), os.path.join(dest, file))
-            print('Moved ' + path + file + ' To Desktop')"""
+            if os.path.exists(path + '/' + file):
+                print('file match: ' + file)
+                shutil.move(os.path.join(path, file), os.path.join(dest, file))
+                print('Moved ' + path + file + ' To FM')
 
     for dir in dirs:
         print('Checking: ' + dir)
@@ -26,9 +26,9 @@ for root, dirs, files in os.walk(path):
                     print(dest + dir + '/' + file)
                     if os.path.exists(dest + dir + '/') == False:
                         os.makedirs(dest + dir + '/')
-                    shutil.move(path + dir + '/' + file, dest + dir + '/' + file)
+                        shutil.move(path + dir + '/' + file, dest + dir + '/' + file)
+
+                    print('Moved ' + path + file + ' To FM')
 
 
-                    
-                    #shutil.move(path + dir + '/' + file, dest + dir +'/' + file)
-                    print('Moved ' + path + file + ' To Desktop')
+
